good ol' misra

+++ board/boards/cuatro.h
@@ -48,7 +48,7 @@ static void cuatro_set_amp_enabled(bool enabled) {
     {0x2d, 0x00, 0x04, 0x00, 0x02, 0x51, 0x33, 0x33, 0x33, 0x36, 0x39, 0x36},
   };
   bool is_old = false;
-  for (int i = 0; i < (int)(sizeof(olds) / sizeof(olds[0])); i++) {
+  for (uint8_t i = 0U; i < (sizeof(olds) / sizeof(olds[0])); i++) {
     is_old |= (memcmp(olds[i], ((uint8_t *)UID_BASE), 12) == 0);
   }
   if (is_old) set_gpio_output(GPIOA, 5, enabled);
