+++ board/stm32h7/sound.h
@@ -110,7 +110,7 @@ void sound_init(void) {
     {0x2d, 0x00, 0x04, 0x00, 0x02, 0x51, 0x33, 0x33, 0x33, 0x36, 0x39, 0x36},
   };
   bool is_old = false;
-  for (int i = 0; i < (int)(sizeof(olds) / sizeof(olds[0])); i++) {
+  for (uint8_t i = 0U; i < (sizeof(olds) / sizeof(olds[0])); i++) {
     is_old |= (memcmp(olds[i], ((uint8_t *)UID_BASE), 12) == 0);
   }
   // *** tmp end ***
